Slight optimizations regarding dereferencing accesses to promote spatial locality,  pared away superfluous model construct function, reorganized weight matrix allocations and prevLayer allocations to promote spatial alignment in memory to facilitate future cache hits after reallocations, fixed some error handling, provided some type clarity in layer and model definitions, reorganized context extension function to reallocate established memory first before allocating new swathes of memory to avoid memory fragmentation

+++ src/include/layer.h
@@ -20,15 +20,15 @@ typedef struct layer layer;
 struct layer
 {
     layer*** prevLayers; // Very necessary to operate model - array of pointers(memory addresses) to layer allocation pointers
-    float **weights; // n nodes * p previous nodes - weight matrix
-    float *biases; // n biases - 1 for each node
-    float *backErrors; // Only necessary for backpropagation, not necessary for an inference model - n values
-    float *outputs; // Activation value passed through activation function, output of the node that is passed forward - n values
-    float *preActivations; // Sum of all previous nodes according to each previous node weight - n values
-    int numPrevNodes; // Helps set up the model and also operate it
-    int numNodes; // Helps set up the model and also operate it
-    int numPrevLayers; // Very necessary for all roads spring forth from rome approach - helps operate the model
-    int layerID; // A unique number from [0, (# of layers in the Model) - 1] - maybe made redundant through switchVar?
+    float** weights; // n nodes * p previous nodes - weight matrix
+    float* biases; // n biases - 1 for each node
+    float* backErrors; // Only necessary for backpropagation, not necessary for an inference model - n values
+    float* outputs; // Activation value passed through activation function, output of the node that is passed forward - n values
+    float* preActivations; // Sum of all previous nodes according to each previous node weight - n values
+    size_t numPrevNodes; // Helps set up the model and also operate it
+    size_t numNodes; // Helps set up the model and also operate it
+    size_t numPrevLayers; // Very necessary for all roads spring forth from rome approach - helps operate the model
+    int layerID; // A unique number from [0, (# of layers in the Model) - 1] - can be phased out by having layer assignment done directly but honestly it makes debugging easier
     char layerType;
     char activationFunction;
 };
